pyproject.toml
[project]
name = "the-truth"
version = "0.1.0"
description = "Use AI to reach the capital-T Truth on any topic, via adversarial debate."
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
truth = "truth.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["truth"]
